[Auth;Biometrics] A lot of security changes
This fixes a lot of security issues on both android and desktop. Now the hash of the key is never stored in plain text in the desktop app. Second, the encryption key is not decryptable without a pin or biometrics. Third, MacOS (probably) should now support biometrics. Also, due to migration from old implementation, the app might need setup again. If you enter the old pin, all the entries should be recovered, else all entries will be wiped